Migracja na Postgres (przyklad53)

+++ przyklad-53/src/main/resources/schema.sql
@@ -2,6 +2,6 @@ CREATE TABLE IF NOT EXISTS airports
 (
     code VARCHAR(3) NOT NULL PRIMARY KEY,
     name VARCHAR(100) NOT NULL,
-    latitude DOUBLE NOT NULL,
-    longitude DOUBLE NOT NULL
+    latitude DOUBLE PRECISION NOT NULL,
+    longitude DOUBLE PRECISION NOT NULL
 );
